chore: changed int to long for id in all files

+++ composeApp/src/commonMain/kotlin/com/example/lootradarkmp/ui/navigation/Screen.kt
@@ -10,6 +10,6 @@ sealed class Screen(val route: String, val label: String, val icon: ImageVector)
     object Notification : Screen("notification", "Notifications", Icons.Default.Notifications)
     object Settings : Screen("settings", "Settings", Icons.Default.Settings)
     object Details : Screen("details/{gameId}", "Game Details", Icons.Default.Info) {
-        fun createRoute(id: Int?) = "details/$id"
+        fun createRoute(id: Long?) = "details/$id"
     }
 }
